Remove equal_range test
The comment stating that std::less requires both types be the same type
is incorrect: std::less<void> serves exactly for this purpose.
This test could not be compiled, because there's no
`operator<(srp::Pack<..>, T)` defined. This operator could be provided
in similar fashion as `operator==`, but std::equal_range would still not
work properly, because one of it's requirement is:

  for all elements, if element < value is true then !(value < element) is also true

which does not hold for srp::Pack in general case.